Configured FluentValidation classes for needed dto's.

+++ PCStore.API/Controllers/AnswerVoteController.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Mvc;
 using PCStore.Application.Features.CQRSDesignPattern.Commands.AnswerVoteCommands;
 using PCStore.Application.Services.AnswerVoteService;
+using System.Security.Claims;
 
 namespace PCStore.API.Controllers
 {
@@ -15,6 +16,10 @@ namespace PCStore.API.Controllers
         [HttpPost]
         public async Task<IActionResult> CreateAnswerVote([FromBody] CreateAnswerVoteCommand request, CancellationToken cancellationToken)
         {
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if (userId is null)
+                return Unauthorized();
+            request.AnswerVoteUserId = userId;
             var result = await _answerVoteService.CreateAnswerVote(request, cancellationToken);
             return StatusCode(result.StatusCode, result.Data);
         }
@@ -23,6 +28,10 @@ namespace PCStore.API.Controllers
         [HttpDelete]
         public async Task<IActionResult> RemoveAnswerVote([FromBody] RemoveAnswerVoteCommand request, CancellationToken cancellationToken)
         {
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if (userId is null)
+                return Unauthorized();
+            request.UserId = userId;
             var result = await _answerVoteService.RemoveAnswerVote(request, cancellationToken);
             return StatusCode(result.StatusCode, result.Message);
         }
