+++ PCStore.API/Controllers/CommentVoteController.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Mvc;
 using PCStore.Application.Features.CQRSDesignPattern.Commands.CommentVoteCommands;
 using PCStore.Application.Services.CommentVoteService;
+using System.Security.Claims;
 
 namespace PCStore.API.Controllers
 {
@@ -15,6 +16,10 @@ namespace PCStore.API.Controllers
         [HttpPost]
         public async Task<IActionResult> CreateCommentVote([FromBody] CreateCommentVoteCommand request, CancellationToken cancellationToken)
         {
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if (userId is null)
+                return Unauthorized();
+            request.CommentVoteUserId = userId;
             var result = await _commentVoteService.CreateCommentVote(request, cancellationToken);
             return StatusCode(result.StatusCode, result.Data);
         }
@@ -23,6 +28,10 @@ namespace PCStore.API.Controllers
         [HttpDelete]
         public async Task<IActionResult> RemoveCommentVote([FromBody] RemoveCommentVoteCommand request, CancellationToken cancellation)
         {
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if (userId is null)
+                return Unauthorized();
+            request.UserId = userId;
             var result = await _commentVoteService.RemoveCommentVote(request, cancellation);
             return StatusCode(result.StatusCode, result);
         }
